requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
Babel==2.3.4
billiard==3.3.0.23
boto==2.45.0
celery==3.1.23
coreapi==1.32.3
defusedxml==0.4.1
Django==1.10.1
django-braces==1.10.0
django-celery==3.1.17
django-cors-middleware==1.3.1
django-filter==0.15.2
django-storages==1.5.2
django-suit==0.2.19
djangorestframework==3.4.1
jet==1.0
jsonfield==2.0.1
drfdocs==0.0.11
easy-thumbnails==2.3
ecdsa==0.13
Fabric==1.12.0
Fabric3==1.11.1.post1
fcm-django==0.2.7
gunicorn==19.6.0
itypes==1.1.0
kombu==3.0.35
oauthlib==1.1.2
openapi-codec==1.0.0
packaging==16.8
paramiko==1.17.2
Pillow==3.3.0
psycopg2==2.6.2
pycrypto==2.6.1
pyfcm==1.2.3
PyJWT==1.4.2
pyparsing==2.2.0
python-social-auth==0.2.21
python3-openid==3.0.10
pytz==2016.6.1
PyYAML==3.12
qrcode==5.3
requests==2.11.0
requests-oauthlib==0.6.2
simplejson==3.8.2
six==1.10.0
uritemplate==3.0.0
urllib3==1.19.1
